requirements.txt
flask
pillow
opencv-python
keras==2.2.5
tensorflow==1.14
h5py
mnist